Removing the tubul_string.h file and using explicit template instantiation to hide the implementation for join.
